+++ cmd/app/dependencies.go
@@ -2,11 +2,13 @@ package app
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"log"
 
 	"github.com/go-git/go-billy/v5/memfs"
 	"github.com/go-git/go-git/v5"
+	"github.com/go-git/go-git/v5/plumbing"
+	"github.com/go-git/go-git/v5/plumbing/format/gitignore"
 	"github.com/go-git/go-git/v5/storage/memory"
 	"github.com/google/go-github/v50/github"
 	"golang.org/x/oauth2"
@@ -17,25 +19,8 @@ import (
 	"github.com/luinunesmeli/goscriba/pkg/config"
 )
 
-func buildGitRepo(cfg config.Config, changelog *tomaster.Changelog) (tomaster.GitRepo, error) {
-	url, err := getRemoteURL(cfg)
-	if err != nil {
-		return tomaster.GitRepo{}, err
-	}
-
-	storer := memory.NewStorage()
-	fs := memfs.New()
-
-	repo, err := git.Clone(storer, fs, &git.CloneOptions{
-		URL:  url,
-		Auth: auth.AuthMethod(cfg),
-	})
-
-	if err != nil {
-		return tomaster.GitRepo{}, fmt.Errorf("actual directory doesn't contains a git repository: %w", err)
-	}
-
-	gitRepo, err := tomaster.NewGitRepo(repo, changelog, cfg, auth.AuthMethod(cfg))
+func buildGitRepo(repo *git.Repository, tree *git.Worktree, cfg config.Config) (tomaster.GitRepo, error) {
+	gitRepo, err := tomaster.NewGitRepo(repo, tree, cfg, auth.AuthMethod(cfg))
 	if err != nil {
 		return tomaster.GitRepo{}, err
 	}
@@ -43,30 +28,50 @@ func buildGitRepo(cfg config.Config, changelog *tomaster.Changelog) (tomaster.Gi
 	return gitRepo, nil
 }
 
-func buildGithubClient(ctx context.Context, cfg config.Config, owner, repo string) tomaster.GithubClient {
+func buildGithub(client *github.Client, cfg config.Config, owner, repo string) tomaster.GithubClient {
+	return tomaster.NewGithubClient(client, cfg, owner, repo)
+}
+
+func githubClient(ctx context.Context, cfg config.Config) *github.Client {
 	ts := oauth2.StaticTokenSource(
 		&oauth2.Token{AccessToken: cfg.GetPersonalAccessToken()},
 	)
 	tc := oauth2.NewClient(ctx, ts)
 
-	return tomaster.NewGithubClient(github.NewClient(tc), cfg, owner, repo)
+	return github.NewClient(tc)
 }
 
-func buildChangelog(cfg config.Config) tomaster.Changelog {
-	return tomaster.NewChangelog(cfg.Changelog)
+func buildChangelog(cfg config.Config, tree *git.Worktree) tomaster.Changelog {
+	return tomaster.NewChangelog(cfg, tree)
 }
 
-func getRemoteURL(cfg config.Config) (string, error) {
-	plainRepo, err := git.PlainOpen(cfg.Path)
+func getWorktree(cfg config.Config, repo *git.Repository) (*git.Worktree, error) {
+	tree, err := repo.Worktree()
 	if err != nil {
-		return "", err
+		return tree, err
 	}
-	repoCfg, err := plainRepo.Config()
-	if err != nil {
-		return "", err
+
+	ignoreList, err := cfg.ReadGitignore()
+	for _, ignore := range ignoreList {
+		tree.Excludes = append(tree.Excludes, gitignore.ParsePattern(ignore, []string{}))
 	}
-	if remotes, ok := repoCfg.Remotes["origin"]; ok && len(remotes.URLs) > 0 {
-		return repoCfg.Remotes["origin"].URLs[0], nil
-	}
-	return "", errors.New("could not load remote URL")
+
+	return tree, err
+}
+
+func cloneRepository(cfg config.Config) (*git.Repository, error) {
+	storer := memory.NewStorage()
+	fs := memfs.New()
+
+	fmt.Println("Loading repository...")
+
+	return git.Clone(storer, fs, &git.CloneOptions{
+		URL:           cfg.Repo.URL,
+		Auth:          auth.AuthMethod(cfg),
+		ReferenceName: plumbing.NewBranchReferenceName("develop"),
+		SingleBranch:  true,
+		NoCheckout:    true,
+		Progress:      log.Writer(),
+		Tags:          git.NoTags,
+	})
 }