+++ tomaster/github.go
@@ -2,12 +2,15 @@ package tomaster
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"log"
 	"net/http"
 
 	"github.com/google/go-github/v50/github"
 
 	"github.com/luinunesmeli/goscriba/pkg/config"
+	"github.com/luinunesmeli/goscriba/pkg/datapool"
 )
 
 type GithubClient struct {
@@ -34,7 +37,7 @@ func NewGithubClient(client *github.Client, cfg config.Config, owner, repo strin
 func (r *GithubClient) LoadLatestTag(ctx context.Context) Task {
 	return Task{
 		Desc: "Loading latest tag",
-		Help: "Couldn't get version. Do you have permission to read this repo?",
+		Help: "Couldn't get version. Do you have permissions to read this repo?",
 		Func: func(session Session) (error, string, Session) {
 			rel, resp, err := r.client.Repositories.GetLatestRelease(ctx, r.owner, r.repo)
 			if err != nil {
@@ -54,7 +57,7 @@ func (r *GithubClient) LoadLatestTag(ctx context.Context) Task {
 func (r *GithubClient) DiffBaseHead(ctx context.Context) Task {
 	return Task{
 		Desc: "Comparing `master` and `develop`",
-		Help: "Couldn't get diff!",
+		Help: "Couldn't resolveq diff!",
 		Func: func(session Session) (error, string, Session) {
 			commits, _, err := r.client.Repositories.CompareCommits(
 				ctx, r.owner, r.repo, r.config.Base, head, &github.ListOptions{},
@@ -63,26 +66,54 @@ func (r *GithubClient) DiffBaseHead(ctx context.Context) Task {
 				return err, "", session
 			}
 
+			cachedCommits := datapool.Pool[string]{}
+			cachedPR := datapool.Pool[int]{}
+
 			session.PRs = PRs{}
 			prOptions := &github.PullRequestListOptions{State: "closed"}
 			for _, commit := range commits.Commits {
+				if cachedCommits.Has(commit.GetSHA()) {
+					log.Println("Hit on commit cache")
+					continue
+				}
+
 				pr, _, _ := r.client.PullRequests.ListPullRequestsWithCommit(ctx, r.owner, r.repo, commit.GetSHA(), prOptions)
 				for _, p := range pr {
+					if cachedPR.Has(p.GetNumber()) {
+						log.Println("PR exists!")
+						continue
+					}
+					cachedPR.Add(p.GetNumber())
+
+					commitsPR, _, _ := r.client.PullRequests.ListCommits(
+						ctx, r.owner, r.repo, p.GetNumber(), &github.ListOptions{},
+					)
+
+					for _, repositoryCommit := range commitsPR {
+						cachedCommits.Add(repositoryCommit.GetSHA())
+					}
+
 					prType := getPRType(p.GetHead())
-					if prType == "" {
+					log.Printf("Number: %d Head: %s Type: %s Merged %t", p.GetNumber(), p.GetHead().GetRef(), prType, p.GetMerged())
+					if !p.GetMerged() && prType == "" {
 						continue
 					}
 
-					session.PRs = session.PRs.Append(PR{
+					session.PRs = append(session.PRs, PR{
 						PRType: prType,
 						Title:  p.GetTitle(),
 						PRLink: p.GetLinks().GetHTML().GetHRef(),
-						Author: commit.GetAuthor().GetLogin(),
+						Author: authorName(commit),
 						Number: p.GetNumber(),
 						Ref:    p.GetHead().GetRef(),
 					})
 				}
 			}
+
+			if len(session.PRs) == 0 {
+				return errors.New("no closed pull requests on `develop` can be merged on `master`"), "", session
+			}
+
 			return nil, "", session
 		},
 	}
@@ -108,8 +139,31 @@ func (r *GithubClient) CreatePullRequest(ctx context.Context) Task {
 			if err != nil {
 				return err, "", session
 			}
+			session.PRUrl = pr.GetHTMLURL()
+			session.PRNumber = pr.GetNumber()
 
 			return nil, fmt.Sprintf("Access at: %s", pr.GetHTMLURL()), session
 		},
+	}
+}
+
+func (r *GithubClient) GetGithubUsername(ctx context.Context) (string, error) {
+	user, _, err := r.client.Users.Get(ctx, "")
+	if err != nil {
+		return "", err
+	}
+	return user.GetLogin(), nil
+}
+
+func authorName(commit *github.RepositoryCommit) string {
+	switch {
+	case commit.GetAuthor().GetLogin() != "":
+		return commit.GetAuthor().GetLogin()
+	case commit.GetCommitter().GetLogin() != "":
+		return commit.GetCommitter().GetLogin()
+	case commit.GetAuthor().GetLogin() != "":
+		return commit.GetAuthor().GetLogin()
+	default:
+		return commit.GetAuthor().GetEmail()
 	}
 }