+++ tomaster/pr.go
@@ -34,6 +34,8 @@ func getPRType(branch *github.PullRequestBranch) PRType {
 		return Feature
 	case strings.HasPrefix(branch.GetRef(), string(Enhancement)):
 		return Enhancement
+	case strings.HasPrefix(branch.GetRef(), string(Fix)):
+		return Fix
 	default:
 		return ""
 	}
@@ -56,21 +58,6 @@ func (p PRs) AsMap() map[PRType]PRs {
 			out[pr.PRType] = PRs{}
 		}
 		out[pr.PRType] = append(out[pr.PRType], pr)
-	}
-	return out
-}
-
-func (p PRs) Append(value PR) PRs {
-	if len(p) == 0 {
-		return PRs{value}
-	}
-
-	out := PRs{}
-	for _, pr := range p {
-		if pr.Number == value.Number {
-			return p
-		}
-		out = append(out, pr)
 	}
 	return out
 }