+++ changelog.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## Version 0.0.3
+**Created at 2023-04-10 by @luinunesmeli**
+
+### Enhancements
+* [Enhancement/more fixes](https://github.com/luinunesmeli/goscriba/pull/87) by @luinunesmeli
+### Fixes
+
 ## Version 0.0.2
 **Created at 2023-04-05 by @luinunesmeli**
 
 ### Enhancements
 * [Remove context timeout, this needs to be addressed properly](https://github.com/luinunesmeli/goscriba/pull/60) by @luinunesmeli
-